requirements.txt
fastapi
uvicorn
langchain
langchain-chroma
chromadb
pdfplumber
python-docx
pytesseract
pdf2image
pandas
python-multipart
requests
aiohttp
spacy
cython
pillow